feat: Небольшие изменения в DTO классах

+++ market-app/src/main/java/ru/yandex/market_app/model/dto/ItemDto.java
@@ -12,5 +12,5 @@ public record ItemDto (
 
     String imgPath,
 
-    Integer count
+    Long count
 ) { }
